fix: Correctly implement automatic consul_key population for --wait consul
Fixed the command line tool to properly populate the consul_key when --wait consul is specified. When --wait consul is used, the tool now sets a placeholder consul_key that will be standardized by the server to services/video-transcription/<job-id>. This resolves the issue where the consul_key wasn't being passed in the API request.